Get address code to a basic test scenario
Get the addressing code to the point where I can test a read and write
into and out of video RAM OK. Memory seems to behave the way I expect it
to. I'll probably streamline this code more as I go, but for now VRAM
works, and by extension most of the rest of low ram.